+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADRoleSetting/MSFT_AADRoleSetting.psm1
@@ -190,6 +190,7 @@ function Get-TargetResource
         -InboundParameters $PSBoundParameters `
         -ProfileName 'beta'
 
+    Select-MgProfile -Name 'beta'
     $MaximumFunctionCount = 32000
 
     Write-Verbose -Message "Getting configuration of Role"
@@ -226,8 +227,19 @@ function Get-TargetResource
     [string[]]$ActivateApprover = @()
     foreach ($Item in $ActivateApprovers.id)
     {
-        $user = Get-MgUser -UserId $Item
-        $ActivateApprover += $user.UserPrincipalName
+        try{
+            $user = Get-MgUser -UserId $Item -ErrorAction Stop
+            $ActivateApprover += $user.UserPrincipalName
+        }
+        catch{
+            try{
+                $group = get-mggroup -GroupId $Item -ErrorAction stop
+                $ActivateApprover += $group.DisplayName
+            }
+            catch{
+                Write-Verbose -Message "Error: $($Error[0])"
+            }
+        }
     }
     $PermanentEligibleAssignmentisExpirationRequired = ($role | Where-Object { $_.Id -eq "Expiration_Admin_Eligibility" }).AdditionalProperties.isExpirationRequired
     $ExpireEligibleAssignment = ($role | Where-Object { $_.Id -eq "Expiration_Admin_Eligibility" }).AdditionalProperties.maximumDuration
@@ -820,14 +832,39 @@ function Set-TargetResource
                 if ($ActivateApprover.count -gt 0)
                 {
                     $primaryApprovers = @()
-                    foreach ($UPN in $ActivateApprover)
-                    {
-                        $ActivateApprovers = @{}
-                        $ActivateApprovers.Add("@odata.type", "#microsoft.graph.singleUser")
-                        $Filter = "UserPrincipalName eq '" + $UPN + "'"
-                        $user = Get-MgUser -Filter $Filter
-                        $ActivateApprovers.Add("userId", $user.Id)
-                        $primaryApprovers += $ActivateApprovers
+                    foreach ($item in $ActivateApprover){
+                        #is not a guid try with user
+                        $Filter = "UserPrincipalName eq '" + $item + "'"
+                        try{
+                            $user = Get-MgUser -Filter $Filter -ErrorAction Stop
+                        }
+                        catch{
+                            Write-Verbose -Message "User not found, try with group"
+                        }
+                        if($user.length -gt 0){
+                            $ActivateApprovers = @{}
+                            $ActivateApprovers.Add("@odata.type", "#microsoft.graph.singleUser")
+                            $ActivateApprovers.Add("userId", $user.Id)
+                            $primaryApprovers += $ActivateApprovers
+                            $user = $null
+                        }
+                        else{
+                            #try with group
+                            $Filter = "Displayname eq '" + $item + "'"
+                            try{
+                                $group = Get-MgGroup -Filter $Filter -ErrorAction Stop
+                            }
+                            catch{
+                                Write-Verbose -Message "Group not found"
+                            }
+                            if($group.length -gt 0){
+                                $ActivateApprovers = @{}
+                                $ActivateApprovers.Add("@odata.type", "#microsoft.graph.groupMembers")
+                                $ActivateApprovers.Add("groupId", $group.Id)
+                                $primaryApprovers += $ActivateApprovers
+                                $group = $null
+                            }
+                        }
                     }
                 }
                 $approvalStages = @{}
@@ -1239,6 +1276,14 @@ function Export-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
+    try{
+        Get-MgPolicyRoleManagementPolicyAssignment -Filter "scopeId eq '/' and scopeType eq 'DirectoryRole'" -ErrorAction Stop
+    }
+    catch{
+        Write-Host ""
+        Write-Host -Message "WARNING: AAD Premium License is required to get the role" -ForegroundColor Yellow
+        continue
+    }
     try
     {
         [array]$roles = Get-MgRoleManagementDirectoryRoleDefinition -Filter $Filter -ErrorAction Stop
@@ -1253,7 +1298,6 @@ function Export-TargetResource
                 ApplicationId         = $ApplicationId
                 TenantId              = $TenantId
                 CertificateThumbprint = $CertificateThumbprint
-                Credential            = $Credential
             }
             $Results = Get-TargetResource @Params
             $Results = Update-M365DSCExportAuthenticationResults -ConnectionMode $ConnectionMode `